Changes as recommended by first-pass code review.
