+++ internal/relay/pinpoint/relay.go
@@ -1,41 +1,13 @@
 package relay
 
 import (
-	"encoding/json"
-	"fmt"
 	"net"
-	"os"
 	"time"
 
 	"github.com/aws/aws-sdk-go/service/pinpointemail"
 	"github.com/aws/aws-sdk-go/service/pinpointemail/pinpointemailiface"
+	"github.com/blueimp/aws-smtp-relay/internal/relay"
 )
-
-// Request object
-type Request struct {
-	Time  time.Time
-	addr  net.Addr
-	IP    string
-	From  string
-	To    []string
-	Error string
-	err   *error
-}
-
-// Log prints the Request data to Stdout/Stderr
-func (req *Request) Log() {
-	req.IP = req.addr.(*net.TCPAddr).IP.String()
-	var out *os.File
-	err := *req.err
-	if err != nil {
-		req.Error = err.Error()
-		out = os.Stderr
-	} else {
-		out = os.Stdout
-	}
-	b, _ := json.Marshal(req)
-	fmt.Fprintln(out, string(b))
-}
 
 // Send uses the given Pinpoint API to send email data
 func Send(
@@ -47,19 +19,18 @@ func Send(
 	setName *string,
 ) {
 	var err error
-	req := &Request{
+	req := &relay.Request{
 		Time: time.Now().UTC(),
-		addr: origin,
+		Addr: origin,
 		From: *from,
 		To:   *to,
-		err:  &err,
+		Err:  &err,
 	}
 	defer req.Log()
 	destinations := []*string{}
 	for k := range *to {
 		destinations = append(destinations, &(*to)[k])
 	}
-
 	_, err = pinpointAPI.SendEmail(&pinpointemail.SendEmailInput{
 		ConfigurationSetName: setName,
 		FromEmailAddress: from,
